+++ scico/flax/examples/examples.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2022-2024 by SCICO Developers
+# Copyright (C) 2022-2025 by SCICO Developers
 # All rights reserved. BSD 3-clause License.
 # This file is part of the SCICO package. Details of the copyright and
 # user license can be found in the 'LICENSE' file distributed with the
@@ -663,7 +663,7 @@ def runtime_error_scalar(
         available: Parameter value available in data.
     """
     raise RuntimeError(
-        f"Requested value of parameter --{type}-- does not match value "
+        f"Requested value of argument '{type}' does not match value "
         f"read from {idstring} file. Requested: {requested}  Available: "
         f"{available}.\nDelete cache and check data source."
     )
@@ -682,7 +682,7 @@ def runtime_error_array(type: str, idstring: str, maxdiff: float):
            entries.
     """
     raise RuntimeError(
-        f"Requested value of parameter --{type}-- does not match value "
+        f"Requested value of argument '{type}' does not match value "
         f"read from {idstring} file. Maximum array difference: "
         f"{maxdiff:>5.3f}.\nDelete cache and check data source."
     )