+++ scico/linop/xray/__init__.py
@@ -66,10 +66,24 @@ CUDA, and can only be run on GPU devices.
 
 import sys
 
-from ._util import rotate_volume
+from ._util import (
+    center_image,
+    image_alignment_rotation,
+    image_centroid,
+    rotate_volume,
+    volume_alignment_rotation,
+)
 from ._xray import XRayTransform2D, XRayTransform3D
 
-__all__ = ["XRayTransform2D", "XRayTransform3D", "rotate_volume"]
+__all__ = [
+    "XRayTransform2D",
+    "XRayTransform3D",
+    "image_centroid",
+    "center_image",
+    "rotate_volume",
+    "image_alignment_rotation",
+    "volume_alignment_rotation",
+]
 
 
 # Imported items in __all__ appear to originate in top-level xray module