+++ scico/optimize/_padmm.py
@@ -167,7 +167,7 @@ class ProximalADMMBase(Optimizer):
             scalar: Current value of the objective function.
         """
         if (x is None) != (z is None):
-            raise ValueError("Both or neither of x and z must be supplied")
+            raise ValueError("Both or neither of arguments 'x' and 'z' must be supplied")
         if x is None:
             x = self.x
             z = self.z
@@ -312,7 +312,7 @@ class ProximalADMM(ProximalADMMBase):
             Norm of primal residual.
         """
         if (x is None) != (z is None):
-            raise ValueError("Both or neither of x and z must be supplied")
+            raise ValueError("Both or neither of arguments 'x' and 'z' must be supplied")
         if x is None:
             x = self.x
             z = self.z
@@ -530,7 +530,7 @@ class NonLinearPADMM(ProximalADMMBase):
             Norm of primal residual.
         """
         if (x is None) != (z is None):
-            raise ValueError("Both or neither of x and z must be supplied")
+            raise ValueError("Both or neither of arguments 'x' and 'z' must be supplied")
         if x is None:
             x = self.x
             z = self.z