+++ CHANGES.rst
@@ -2,8 +2,14 @@
 SCICO Release Notes
 ===================
 
+Version 0.0.8   (unreleased)
+----------------------------
 
-Version 0.0.7   (unreleased)
+• No changes yet.
+
+
+
+Version 0.0.7   (2025-12-09)
 ----------------------------
 
 • New module ``scico.trace`` for tracing function/method calls.
@@ -11,8 +17,16 @@ Version 0.0.7   (unreleased)
   a functional composed with an orthogonal linear operator.
 • New optimizer methods ``save_state`` and ``load_state`` supporting
   algorithm state checkpointing.
-• Support ``jaxlib`` and ``jax`` versions 0.4.13 to 0.6.2.
-• Support ``flax`` versions 0.8.0 to 0.10.6.
+• New classes for creating a volume from an image by symmetry, and
+  for cone beam X-ray transform of a cylindrically symmetric object
+  in module ``linop.xray.symcone``.
+• New utility functions for CT reconstruction preprocessing added in
+  module ``linop.xray``.
+• Moved ``linop.abel`` module to ``linop.xray.abel``.
+• Make ``orbax-checkpoint`` dependency optional due to absence of recent
+  conda-forge packages.
+• Support ``jaxlib`` and ``jax`` versions 0.5.0 to 0.8.1.
+• Support ``flax`` versions 0.8.0 to 0.12.0